% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{accuracy}
\alias{accuracy}
\title{This function evaluates the performance of different techniques}
\usage{
accuracy(pred, actual)
}
\description{
This function evaluates the performance of different techniques
}
