% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortSeperator.R
\name{cohortSeperator}
\alias{cohortSeperator}
\title{Generate a different dataframe for every cohort year}
\usage{
cohortSeperator(Range, CohortCol, YearCol, RetentionStatus, dataframe, interval)
}
\arguments{
\item{Range}{Range of years in dataset}

\item{CohortCol}{Column which contains the indentifier of the cohort}

\item{YearCol}{Column which contains the year}

\item{RetentionStatus}{The column which contains the retention status of the individual}

\item{dataframe}{The data itself}

\item{interval}{The distance from the cohort year you'd like to look at. (ex. look at the retention status for
those cohort year 2008 in the year 2012, interval = 4)}
}
\description{
Not totally useful, but there if you need it.
}
\examples{
cohortSeperator(2008:2014,
                         CohortCol = "Cohort Year",
                         RetentionStatus = "RetentionStatus",
                         dat, interval = 1)
}
