% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnCollapse.R
\name{columnCollapse}
\alias{columnCollapse}
\title{Collapse responses from a multiple response into one column}
\usage{
columnCollapse(columns, colname, dataframe)
}
\arguments{
\item{columns}{columns which contain a string to be collapse}

\item{colname}{column name to output to}

\item{dataframe}{Dataframe to be collapsed}
}
\description{
Collapse responses from a multiple response into one column
}
\examples{
columnCollapse("city moved to", "cities moved to", GradSurvey)

}
