% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedded_fs.R
\name{emb_mod}
\alias{emb_mod}
\title{This function runs embedded model}
\usage{
emb_mod(
  df = traindf,
  outvar = "y",
  modeltype = "regression",
  technique = "rf",
  ...
)
}
\description{
This function runs embedded model
}
