% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_fs.R
\name{ensem_homo}
\alias{ensem_homo}
\title{This function runs ensemble homogenous model}
\usage{
ensem_homo(
  df = traindf,
  outvar = "y",
  modeltype = "regression",
  technique = "rf",
  sampletype = "bootsample",
  fstype = "randomfeat",
  maxfeatsample = ncol(traindf) - 2,
  boots = 100,
  parallel = T,
  perfparallel = T,
  Trace = T,
  savememory = T,
  smart = T,
  ...
)
}
\description{
This function runs ensemble homogenous model
}
