% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fs.R
\name{filter_mod}
\alias{filter_mod}
\title{This function runs filter model}
\usage{
filter_mod(
  df = traindf,
  outvar = "y",
  modeltype = "regression",
  technique = "uva",
  ...
)
}
\description{
This function runs filter model
}
