% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foip.R
\name{foip}
\alias{foip}
\title{FOIP your data!}
\usage{
foip(column, dataframe)
}
\arguments{
\item{column}{column to subset on}

\item{dataframe}{data to be foiped}
}
\description{
Really can be useful for protecting privacy, but have to be aware of how to use it safely.
}
\examples{
foip(program, survey)

}
