% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_clean.R
\name{fuzzy_clean}
\alias{fuzzy_clean}
\title{Use clean fuzzy strings by using approximated string matching}
\usage{
fuzzy_clean(vector, distance, iterations)
}
\arguments{
\item{vector}{Vector containing strings to clean}

\item{distance}{Distance between strings to match.
Can take a ratio or a whole number.
May also take a list as an argument as per agrep().}

\item{iterations}{Number of itereations the function will do. A smart idea is to tune the function with low distance and high iterations.}
}
\description{
Use clean fuzzy strings by using approximated string matching
}
\examples{
df$Vector2 <- fuzzy_clean(df$Vector1, 0.25, 5)
}
