% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameter_function.R
\name{hyperaparameter}
\alias{hyperaparameter}
\title{Hyperparameter optimization functions}
\usage{
hyperaparameter(
  df = traindf,
  outvar = "y",
  modeltype = "regression",
  technique = "rf",
  ...
)
}
\description{
These function optimize the different techniques for Survival, Binary or Continuous
}
