% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{inputdataset}
\alias{inputdataset}
\title{Here functions generate artificial data for regression, logistic and survival}
\usage{
inputdataset(
  varnum,
  setting = "No_Correlation",
  seed = 2,
  correlation_var = 15,
  correlation_val = 5,
  train_sample = 500,
  datadist = c("random", "normal"),
  corr = c("fixed", "random")
)
}
\description{
Here functions generate artificial data for regression, logistic and survival
}
