% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{modelanalysis}
\alias{modelanalysis}
\title{This function evaluates the different models used in GA function}
\usage{
modelanalysis(df_train, df_test, outvar, modeltype, technique, hyperpara = NA)
}
\description{
This function evaluates the different models used in GA function
}
