% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{modelfunction}
\alias{modelfunction}
\title{Support functions}
\usage{
modelfunction(seed = 1, traindf, outvar, modeltype, technique, hyperpara = NA)
}
\description{
This function run the different techniques for Survival, Binary or Continuous
}
