% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleResponseCountList.R
\name{multipleResponseCountList}
\alias{multipleResponseCountList}
\title{Count the number of occurances of strings in a column}
\usage{
multipleResponseCountList(list, column, data)
}
\arguments{
\item{list}{The list of strings you want to count and put into dataframe}

\item{column}{The columnname you'd like to search}

\item{data}{Dataset to be searched}
}
\description{
Not super useful, but there when you need it.
}
\examples{
multipleResponseCountList(c("medicine hat", "lethbridge", "calgary"),
         "Cities", "Bathroom Habits", "Bathroom Use Count")
}
