% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_technique.R
\name{penaltytech}
\alias{penaltytech}
\title{Support functions}
\usage{
penaltytech(df, outvar, modeltype = "regression", ...)
}
\description{
This function define the different techniques for Survival, Binary or Continuous
}
