% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{predperf}
\alias{predperf}
\title{Performance Functions}
\usage{
predperf(
  traindf,
  testdf = NA,
  outvar,
  modeltype = "regression",
  martarget = NA,
  selfeat = NA,
  pretech,
  cv_split = 30,
  savememory = F,
  Trace = T,
  parallel = F,
  smart = F,
  perfparallel = F,
  interaction = T
)
}
\description{
Performance Functions
}
