% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_aggregation.R
\name{rankaggfun}
\alias{rankaggfun}
\title{This function runs rank aggregation}
\usage{
rankaggfun(
  rankmethod,
  model_list = NA,
  techlist = NA,
  ranklist = NA,
  covs = NA,
  ...
)
}
\description{
This function runs rank aggregation
}
