% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round a number. There is this stupid issue with round() that rounds .5 down..floating point nonsense.}
\usage{
round2(x, n)
}
\arguments{
\item{x}{Number you'd like to round}

\item{digit}{Number of digits you'd like to round}
}
\description{
Round a number. There is this stupid issue with round() that rounds .5 down..floating point nonsense.
}
\examples{
round2(x,2)

}
