% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.R
\name{SDI}
\alias{SDI}
\title{Generate a Simpson Diversity Index Number}
\usage{
SDI(col, data, type)
}
\arguments{
\item{data}{Dataset}

\item{type}{The type of Simpson index you'd like to generate. Reciprocal is 1/Pa, Dominance is 1-Pa, Blank is just the Simpson Index}

\item{column}{Column which contains the count/tally of people from each country.}
}
\description{
Generate a Simpson Diversity Index Number
}
\examples{
SDI("Unique Student Static", EnrolmentData, type = "reciprocal")

}
