% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smry.R
\name{smry}
\alias{smry}
\title{Wrapper for summary(as.factor(Vec))}
\usage{
smry(vector)
}
\arguments{
\item{vector}{column to subset on}
}
\description{
Just provides a simple frequency table of a vector
}
\examples{
smry(df$Program)

}
