\name{sraMakeObject}
\alias{sraMakeObject}

\title{Creates an object of class "srafit" (internal function) }
\description{
  The content of an object of class \code{srafit} correponding to several models is very similar; this function initializes a new object and performs basic instructions.
}
\usage{
sraMakeObject(sradata, model, start, fixed, FUNtimeseries)
}

\arguments{
  \item{sradata}{ A data object generated by \code{\link{sraData}}. }
  \item{model}{ The result of the call of \code{\link{mle}}. }
  \item{start}{ The list of the starting values (for the function to know which parameters were active). }
  \item{fixed}{ The list of the fixed values. }
  \item{FUNtimeseries}{ The time series function, one among those listed in \code{\link{sraAutoregTimeseries}}. }
}

\value{
  An object of class \code{srafit} (see \code{\link{sraAutoreg}} for more details).
}

\note{ This function should not be called by the end user. }
\seealso{ \code{\link{sraAutoreg}}, \code{\link{sraCstvar}}, and other mechanistic models, \code{\link{sraAutoregTimeseries}}. }


\keyword{ internal }

