% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringFilter.r
\name{stringFilter}
\alias{stringFilter}
\title{Filter cells containing a string from a dataset}
\usage{
stringFilter(list, column, data)
}
\arguments{
\item{list}{List of strings you'd like to filter out of a dataset}

\item{column}{Column you'd like to search for the strings in}

\item{data}{Data you'd like to filter}
}
\description{
Filter cells containing a string from a dataset
}
\examples{
stringFilter(list, column, data)

}
