% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringSubset.R
\name{stringSubset}
\alias{stringSubset}
\title{Subset a dataframe by a string}
\usage{
stringSubset(list, column, data)
}
\arguments{
\item{list}{List of strings you'd like to subset a dataset by}

\item{column}{Column you'd like to search for the strings in}

\item{data}{Data you'd like to subset}
}
\description{
Use when you want to use a regex search to filter an element by a string.
Can be slow using regex, so try to avoid using it inside loops.
}
\examples{
stringSubset(list, column, data)

}
