% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmR.R
\name{xmR}
\alias{xmR}
\title{Generate the XMR data for any time-series data. BUT DO NOT USE THIS}
\usage{
xmR(df, measure, interval)
}
\arguments{
\item{df}{The dataframe containing the time-series data. Must be in long format.
At least 1 variable for time and one variable for measure.}

\item{measure}{The column containing the measure. Must be in numeric format.}

\item{interval}{The interval you'd like to use to calculate the averages. Defaults to 5.}
}
\description{
Will be used to calculate XMR data.
}
\examples{
dat.xmr <- xmR(dat, "Measure", 5)
dat.xmr <- dat \%>\% group_by(., Program, Variable) \%>\% do(xmR2(., "Measure"))

}
