% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmR_chart.R
\name{xmR_chart}
\alias{xmR_chart}
\title{Generate the XMR chart for XMR data}
\usage{
xmR_chart(dataframe, time, measure, facetvar)
}
\arguments{
\item{dataframe}{Output from xmR()}

\item{measure}{Measure}

\item{facetvar}{The column containing the factor you'd like to facet on.}
}
\description{
Useful for diagnostics on xmR, and just visualizing the data.
}
\examples{
dat.xmr <- xmR(dat, "Measure", 5)
dat.xmr <- dat \%>\% 
            group_by(., Program, Variable) \%>\% 
            do(xmR(., "Measure")) \%>\% 
            do(xmR_Chart(., "Variable"))

}
